Refactored imagewithlocation into separate file
TODO: rationalise the stuff in utils…
